+++ CHANGELOG.md
@@ -8,6 +8,15 @@
 - Added `.gitignore`
 - Added `Project` model
 - Added `Task` model
+- Registered the `Project` and `Task` models in admin
+
+### Changed
+- Updated the `ProjectAdmin` to better display the projects in the admin panel
+- Moved all choices to `constants.py`
+- Updated the choices for task status
+
+### Fixed
+- Resolved the constants import error
 
 ## [Setup]
 - Setup project